refactor: redistribute methods by module
