+++ iamport/__init__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
 from .client import Iamport
+from . import exceptions
 
-__all__ = ['Iamport']
+__all__ = ['Iamport', 'exceptions']
+
