+++ tests/test_pay_unschedule.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 import time
 
+from iamport.exceptions import ResponseError
+
 
 def test_pay_unschedule(iamport):
     payload_without_customer_uid = {
@@ -20,6 +22,6 @@ def test_pay_unschedule(iamport):
 
     try:
         iamport.pay_unschedule(**payload_full)
-    except iamport.ResponseError as e:
+    except ResponseError as e:
         assert e.code == 1
         assert u'취소할 예약결제 기록이 존재하지 않습니다.' in e.message
